大致完成所有接口
